apps/backend/requirements.txt
# CryptoAuditAI Backend Dependencies

# Core FastAPI and async support
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
asyncpg==0.29.0  # For PostgreSQL support if needed
alembic==1.12.1  # Database migrations

# Security and encryption
bcrypt==4.1.1
cryptography==41.0.8
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic-settings==2.1.0

# AI and ML components
sentence-transformers==2.2.2
numpy==1.24.4
scikit-learn==1.3.2
faiss-cpu==1.7.4  # Vector similarity search
torch==2.1.1
transformers==4.36.2

# Blockchain integrations
web3==6.11.3
ccxt==4.1.49
eth-account==0.9.0
eth-utils==2.3.0

# HTTP client and web scraping
httpx==0.25.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
requests==2.31.0

# Report generation
reportlab==4.0.7
matplotlib==3.8.2
plotly==5.17.0
jinja2==3.1.2

# Data processing
pandas==2.1.4
openpyxl==3.1.2

# Logging and monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Production deployment
gunicorn==21.2.0
redis==5.0.1  # For caching and task queues if needed
celery==5.3.4  # Background task processing

# Additional utilities
click==8.1.7
python-dateutil==2.8.2
pytz==2023.3
pydantic==2.5.1